# gabbler-desc/src/main/resources/openapi/api/gabbler-api.yaml
openapi: 3.0.3
info:
  title: "Documentation concernant l'api gabbler "
  description: "Documentation concernant l'api gabbler "
  version: 1.0.0
servers:
  - url: /api
    description: serveur de l'api
paths:
  /doc/json:
    get:
      summary: Récupère le json doc
      description: Récupère le json doc
      operationId: getJsonDoc
      responses:
        '200':
          description: Retourne la string json pour la doc avec succès.
          content:
            application/json:
              schema:
                $ref: '../model/gabbler-model.yaml#/components/schemas/JsonDoc'
        '400':
          description: Erreur retournée quand il y a un problème avec les paramètres.
          content:
            application/json:
              schema:
                $ref: '../model/gabbler-model.yaml#/components/schemas/AccessError'
        '403':
          description: Erreur retournée quand l'appelant n'est pas autorisé à réaliser cette action (i.e. ressource ne lui appartenant pas).
          content:
            application/json:
              schema:
                $ref: '../model/gabbler-model.yaml#/components/schemas/AccessError'
        '404':
          description: Erreur retournée quand la ressource demandée n'a pas été trouvée.
          content:
            application/json:
              schema:
                $ref: '../model/gabbler-model.yaml#/components/schemas/AccessError'
        '500':
          description: Erreur retournée quand une erreur non maitrisée a été rencontrée par l'API.
          content:
            application/json:
              schema:
                $ref: '../model/gabbler-model.yaml#/components/schemas/AccessError'
        '503':
          description: Erreur retournée quand l'un des services utilisés par l'API n'est pas disponible.
          content:
            application/json:
              schema:
                $ref: '../model/gabbler-model.yaml#/components/schemas/AccessError'
      tags:
        - doc

  /users:
    get:
      summary: Récupère la liste des utilisateurs
      description: Récupère la liste des utilisateurs
      operationId: getUsers
      security:
        - bearerAuth: [ ]
      responses:
        '200':
          description: Retourne la liste des utilisateurs avec succès.
          content:
            application/json:
              schema:
                items:
                  $ref: '../model/gabbler-model.yaml#/components/schemas/User'
        '400':
          description: Erreur retournée quand il y a un problème avec les paramètres.
          content:
            application/json:
              schema:
                $ref: '../model/gabbler-model.yaml#/components/schemas/AccessError'
        '403':
          description: Erreur retournée quand l'appelant n'est pas autorisé à réaliser cette action (i.e. ressource ne lui appartenant pas).
          content:
            application/json:
              schema:
                $ref: '../model/gabbler-model.yaml#/components/schemas/AccessError'
        '404':
          description: Erreur retournée quand la ressource demandée n'a pas été trouvée.
          content:
            application/json:
              schema:
                $ref: '../model/gabbler-model.yaml#/components/schemas/AccessError'
        '500':
          description: Erreur retournée quand une erreur non maitrisée a été rencontrée par l'API.
          content:
            application/json:
              schema:
                $ref: '../model/gabbler-model.yaml#/components/schemas/AccessError'
        '503':
          description: Erreur retournée quand l'un des services utilisés par l'API n'est pas disponible.
          content:
            application/json:
              schema:
                $ref: '../model/gabbler-model.yaml#/components/schemas/AccessError'

      tags:
        - user

  /users/{uuid}:
    get:
      summary: Récupère un utilisateur par son uuid
      description: Récupère un utilisateur par son uuid
      operationId: getUserById
      security:
        - bearerAuth: [ ]
      parameters:
        - name: uuid
          in: path
          description: uuid de l'utilisateur
          required: true
          schema:
            type: string
      responses:
        '200':
          description: Retourne la liste des utilisateurs avec succès.
          content:
            application/json:
              schema:
                $ref: '../model/gabbler-model.yaml#/components/schemas/User'
        '400':
          description: Erreur retournée quand il y a un problème avec les paramètres.
          content:
            application/json:
              schema:
                $ref: '../model/gabbler-model.yaml#/components/schemas/AccessError'
        '403':
          description: Erreur retournée quand l'appelant n'est pas autorisé à réaliser cette action (i.e. ressource ne lui appartenant pas).
          content:
            application/json:
              schema:
                $ref: '../model/gabbler-model.yaml#/components/schemas/AccessError'
        '404':
          description: Erreur retournée quand la ressource demandée n'a pas été trouvée.
          content:
            application/json:
              schema:
                $ref: '../model/gabbler-model.yaml#/components/schemas/AccessError'
        '500':
          description: Erreur retournée quand une erreur non maitrisée a été rencontrée par l'API.
          content:
            application/json:
              schema:
                $ref: '../model/gabbler-model.yaml#/components/schemas/AccessError'
        '503':
          description: Erreur retournée quand l'un des services utilisés par l'API n'est pas disponible.
          content:
            application/json:
              schema:
                $ref: '../model/gabbler-model.yaml#/components/schemas/AccessError'
      tags:
        - user

  /gabs:
    get:
      summary: Récupère la liste des gabs
      description: Récupère la liste des gabs
      operationId: getGabs
      security:
        - bearerAuth: [ ]
      responses:
        '200':
          description: Retourne la liste des gabs avec succès.
          content:
            application/json:
              schema:
                items:
                  $ref: '../model/gabbler-model.yaml#/components/schemas/Gab'
        '400':
          description: Erreur retournée quand il y a un problème avec les paramètres.
          content:
            application/json:
              schema:
                $ref: '../model/gabbler-model.yaml#/components/schemas/AccessError'
        '403':
          description: Erreur retournée quand l'appelant n'est pas autorisé à réaliser cette action (i.e. ressource ne lui appartenant pas).
          content:
            application/json:
              schema:
                $ref: '../model/gabbler-model.yaml#/components/schemas/AccessError'
        '404':
          description: Erreur retournée quand la ressource demandée n'a pas été trouvée.
          content:
            application/json:
              schema:
                $ref: '../model/gabbler-model.yaml#/components/schemas/AccessError'
        '500':
          description: Erreur retournée quand une erreur non maitrisée a été rencontrée par l'API.
          content:
            application/json:
              schema:
                $ref: '../model/gabbler-model.yaml#/components/schemas/AccessError'
        '503':
          description: Erreur retournée quand l'un des services utilisés par l'API n'est pas disponible.
          content:
            application/json:
              schema:
                $ref: '../model/gabbler-model.yaml#/components/schemas/AccessError'
      tags:
        - gab

  /gabs/{id}:
    get:
      summary: Récupère un gab par son id
      description: Récupère un gab par son id
      operationId: getGabById
      security:
        - bearerAuth: [ ]
      parameters:
        - name: id
          in: path
          description: id du gab
          required: true
          schema:
            type: integer
      responses:
        '200':
          description: Retourne une gab avec succès.
          content:
            application/json:
              schema:
                $ref: '../model/gabbler-model.yaml#/components/schemas/Gab'
        '400':
          description: Erreur retournée quand il y a un problème avec les paramètres.
          content:
            application/json:
              schema:
                $ref: '../model/gabbler-model.yaml#/components/schemas/AccessError'
        '403':
          description: Erreur retournée quand l'appelant n'est pas autorisé à réaliser cette action (i.e. ressource ne lui appartenant pas).
          content:
            application/json:
              schema:
                $ref: '../model/gabbler-model.yaml#/components/schemas/AccessError'
        '404':
          description: Erreur retournée quand la ressource demandée n'a pas été trouvée.
          content:
            application/json:
              schema:
                $ref: '../model/gabbler-model.yaml#/components/schemas/AccessError'
        '500':
          description: Erreur retournée quand une erreur non maitrisée a été rencontrée par l'API.
          content:
            application/json:
              schema:
                $ref: '../model/gabbler-model.yaml#/components/schemas/AccessError'
        '503':
          description: Erreur retournée quand l'un des services utilisés par l'API n'est pas disponible.
          content:
            application/json:
              schema:
                $ref: '../model/gabbler-model.yaml#/components/schemas/AccessError'
      tags:
        - gab
  /gabs/new/{uuid_user}:
    post:
      summary: Créer un gab
      description: Créer un gab
      operationId: createGab
      security:
        - bearerAuth: [ ]
      parameters:
        - name: uuid_user
          in: path
          description: uuid de l'utilisateur
          required: true
          schema:
            type: string
      requestBody:
        description: Gab à créer
        required: true
        content:
          application/json:
            schema:
              $ref: '../model/gabbler-model.yaml#/components/schemas/GabCreation'
      responses:
        '200':
          description: Gab créé avec succès.
          content:
            application/string:
              schema:
                type: string

        '400':
          description: Erreur retournée quand il y a un problème avec les paramètres.
          content:
            application/json:
              schema:
                $ref: '../model/gabbler-model.yaml#/components/schemas/AccessError'
        '403':
          description: Erreur retournée quand l'appelant n'est pas autorisé à réaliser cette action (i.e. ressource ne lui appartenant pas).
          content:
            application/json:
              schema:
                $ref: '../model/gabbler-model.yaml#/components/schemas/AccessError'
        '404':
          description: Erreur retournée quand la ressource demandée n'a pas été trouvée.
          content:
            application/json:
              schema:
                $ref: '../model/gabbler-model.yaml#/components/schemas/AccessError'
        '500':
          description: Erreur retournée quand une erreur non maitrisée a été rencontrée par l'API.
          content:
            application/json:
              schema:
                $ref: '../model/gabbler-model.yaml#/components/schemas/AccessError'
        '503':
          description: Erreur retournée quand l'un des services utilisés par l'API n'est pas disponible.
          content:
            application/json:
              schema:
                $ref: '../model/gabbler-model.yaml#/components/schemas/AccessError'
      tags:
        - gab

  /gabs/{id}/comments:
    get:
      summary: Récupère la liste des commentaires d'un gab
      description: Récupère la liste des commentaires d'un gab
      operationId: getCommentsByGabId
      security:
        - bearerAuth: [ ]
      parameters:
        - name: id
          in: path
          description: id du gab
          required: true
          schema:
            type: integer
      responses:
        '200':
          description: Retourne la liste des commentaires d'un gab avec succès.
          content:
            application/json:
              schema:
                items:
                  $ref: '../model/gabbler-model.yaml#/components/schemas/Gab'
        '400':
          description: Erreur retournée quand il y a un problème avec les paramètres.
          content:
            application/json:
              schema:
                $ref: '../model/gabbler-model.yaml#/components/schemas/AccessError'
        '403':
          description: Erreur retournée quand l'appelant n'est pas autorisé à réaliser cette action (i.e. ressource ne lui appartenant pas).
          content:
            application/json:
              schema:
                $ref: '../model/gabbler-model.yaml#/components/schemas/AccessError'
        '404':
          description: Erreur retournée quand la ressource demandée n'a pas été trouvée.
          content:
            application/json:
              schema:
                $ref: '../model/gabbler-model.yaml#/components/schemas/AccessError'
        '500':
          description: Erreur retournée quand une erreur non maitrisée a été rencontrée par l'API.
          content:
            application/json:
              schema:
                $ref: '../model/gabbler-model.yaml#/components/schemas/AccessError'
        '503':
          description: Erreur retournée quand l'un des services utilisés par l'API n'est pas disponible.
          content:
            application/json:
              schema:
                $ref: '../model/gabbler-model.yaml#/components/schemas/AccessError'
      tags:
        - gab

  /gabs/feed:
    get:
      operationId: getFeed
      summary: Récupère le feed
      description: Récupère le feed
      security:
        - bearerAuth: [ ]
      responses:
        '200':
          description: Retourne la liste des gabs triée par le plus récent au moins récent correspondant au feed.
          content:
            application/json:
              schema:
                items:
                  $ref: '../model/gabbler-model.yaml#/components/schemas/Gab'
        '400':
          description: Erreur retournée quand il y a un problème avec les paramètres.
          content:
            application/json:
              schema:
                $ref: '../model/gabbler-model.yaml#/components/schemas/AccessError'
        '403':
          description: Erreur retournée quand l'appelant n'est pas autorisé à réaliser cette action (i.e. ressource ne lui appartenant pas).
          content:
            application/json:
              schema:
                $ref: '../model/gabbler-model.yaml#/components/schemas/AccessError'
        '404':
          description: Erreur retournée quand la ressource demandée n'a pas été trouvée.
          content:
            application/json:
              schema:
                $ref: '../model/gabbler-model.yaml#/components/schemas/AccessError'
        '500':
          description: Erreur retournée quand une erreur non maitrisée a été rencontrée par l'API.
          content:
            application/json:
              schema:
                $ref: '../model/gabbler-model.yaml#/components/schemas/AccessError'
        '503':
          description: Erreur retournée quand l'un des services utilisés par l'API n'est pas disponible.
          content:
            application/json:
              schema:
                $ref: '../model/gabbler-model.yaml#/components/schemas/AccessError'
      tags:
        - gab

  /authenticate:
    post:
      operationId: authenticateAndGetToken
      requestBody:
        description: credentials de l'utilisateur
        required: true
        content:
          application/json:
            schema:
              $ref: '../model/gabbler-model.yaml#/components/schemas/UserAuth'
      responses:
        '200':
          description: Retourne l'utilisateur et son token avec succès.
          content:
            application/json:
              schema:
                $ref: '../model/gabbler-model.yaml#/components/schemas/UserToken'
        '400':
          description: Erreur retournée quand il y a un problème avec les paramètres.
          content:
            application/json:
              schema:
                $ref: '../model/gabbler-model.yaml#/components/schemas/AccessError'
        '403':
          description: Erreur retournée quand l'appelant n'est pas autorisé à réaliser cette action (i.e. ressource ne lui appartenant pas).
          content:
            application/json:
              schema:
                $ref: '../model/gabbler-model.yaml#/components/schemas/AccessError'
        '404':
          description: Erreur retournée quand la ressource demandée n'a pas été trouvée.
          content:
            application/json:
              schema:
                $ref: '../model/gabbler-model.yaml#/components/schemas/AccessError'
        '500':
          description: Erreur retournée quand une erreur non maitrisée a été rencontrée par l'API.
          content:
            application/json:
              schema:
                $ref: '../model/gabbler-model.yaml#/components/schemas/AccessError'
        '503':
          description: Erreur retournée quand l'un des services utilisés par l'API n'est pas disponible.
          content:
            application/json:
              schema:
                $ref: '../model/gabbler-model.yaml#/components/schemas/AccessError'
      tags:
        - auth
  /register:
    post:
      operationId: registerUser
      requestBody:
        description: credentials de l'utilisateur
        required: true
        content:
          application/json:
            schema:
              $ref: '../model/gabbler-model.yaml#/components/schemas/UserRegister'
      responses:
        '200':
          description: Retourne l'utilisateur et son token avec succès.
          content:
            application/json:
              schema:
                $ref: '../model/gabbler-model.yaml#/components/schemas/UserToken'
        '400':
          description: Erreur retournée quand il y a un problème avec les paramètres.
          content:
            application/json:
              schema:
                $ref: '../model/gabbler-model.yaml#/components/schemas/AccessError'
        '403':
          description: Erreur retournée quand l'appelant n'est pas autorisé à réaliser cette action (i.e. ressource ne lui appartenant pas).
          content:
            application/json:
              schema:
                $ref: '../model/gabbler-model.yaml#/components/schemas/AccessError'
        '404':
          description: Erreur retournée quand la ressource demandée n'a pas été trouvée.
          content:
            application/json:
              schema:
                $ref: '../model/gabbler-model.yaml#/components/schemas/AccessError'
        '500':
          description: Erreur retournée quand une erreur non maitrisée a été rencontrée par l'API.
          content:
            application/json:
              schema:
                $ref: '../model/gabbler-model.yaml#/components/schemas/AccessError'
        '503':
          description: Erreur retournée quand l'un des services utilisés par l'API n'est pas disponible.
          content:
            application/json:
              schema:
                $ref: '../model/gabbler-model.yaml#/components/schemas/AccessError'
      tags:
        - auth

  /search:
    get:
      operationId: searchUser
      summary: Recherche un utilisateur
      description: Recherche un utilisateur
      security:
        - bearerAuth: [ ]
      parameters:
        - name: content
          in: query
          description: Contenu de la recherche
          required: true
          schema:
            type: string
      responses:
        '200':
          description: Retourne les utilisateurs et gabs correspondants à la recherche.
          content:
            application/json:
              schema:
                $ref: '../model/gabbler-model.yaml#/components/schemas/SearchContent'
        '400':
          description: Erreur retournée quand il y a un problème avec les paramètres.
          content:
            application/json:
              schema:
                $ref: '../model/gabbler-model.yaml#/components/schemas/AccessError'
        '403':
          description: Erreur retournée quand l'appelant n'est pas autorisé à réaliser cette action (i.e. ressource ne lui appartenant pas).
          content:
            application/json:
              schema:
                $ref: '../model/gabbler-model.yaml#/components/schemas/AccessError'
        '404':
          description: Erreur retournée quand la ressource demandée n'a pas été trouvée.
          content:
            application/json:
              schema:
                $ref: '../model/gabbler-model.yaml#/components/schemas/AccessError'
        '500':
          description: Erreur retournée quand une erreur non maitrisée a été rencontrée par l'API.
          content:
            application/json:
              schema:
                $ref: '../model/gabbler-model.yaml#/components/schemas/AccessError'
        '503':
          description: Erreur retournée quand l'un des services utilisés par l'API n'est pas disponible.
          content:
            application/json:
              schema:
                $ref: '../model/gabbler-model.yaml#/components/schemas/AccessError'
      tags:
        - search